Function for Appending multiple controls at once
Append() adds one control at a time, while it is a common
use-case that a user wants to add several controls. This
leads to the unnecessary repetion of a single function call.

AppendMultiple() appends controls as non-stretchy, while
AppendMultipleEx() allows boolean to explicitely specify
the type of control.

Standard Append() was changed to use AppendMultipleEx() too.